+++ src/melee/ft/chara/ftMario/ftMr_SpecialLw.c
@@ -108,7 +108,7 @@ void ftMr_SpecialAirLw_Enter(HSD_GObj* gobj)
     } else {
         sub_val = sa->speciallw.tap_y_vel_max;
     }
-    fp->self_vel.y = (float) (sa->speciallw.vel_y - sub_val);
+    fp->self_vel.y = (sa->speciallw.vel_y - sub_val);
     ftCommon_ClampSelfVelX(fp, sa->speciallw.air_momentum_x);
     doStartMotion(gobj);
     fp->pre_hitlag_cb = &efLib_PauseAll;
@@ -138,7 +138,7 @@ void ftMr_SpecialAirLw_Anim(HSD_GObj* gobj)
     Fighter* fp = GET_FIGHTER(gobj);
     ftMario_DatAttrs* sa = (ftMario_DatAttrs*) fp->dat_attrs;
 
-    if ((u32) fp->cmd_vars[1] != 0U) {
+    if (fp->cmd_vars[1] != 0U) {
         fp->cmd_vars[1] = 0U;
         fp->u.mr.x2234_tornadoCharge = 1;
     }
@@ -217,7 +217,7 @@ void ftMr_SpecialAirLw_Phys(HSD_GObj* gobj)
     sa = fp->dat_attrs;
 
     if (((s32) fp->u.mr.x2234_tornadoCharge == false) &&
-        ((u32) fp->cmd_vars[2] != 0U) && ((fp->input.x668 & HSD_PAD_B) != 0))
+        (fp->cmd_vars[2] != 0U) && ((fp->input.x668 & HSD_PAD_B) != 0))
     {
         ftCommon_Ascend(fp, sa->speciallw.tap_y_vel_max,
                         sa->speciallw.tap_grav);
@@ -225,10 +225,9 @@ void ftMr_SpecialAirLw_Phys(HSD_GObj* gobj)
     ftCommon_FallBasic(fp);
     flt_var = sa->speciallw.air_momentum_x;
     sa_2 = fp->dat_attrs;
-    if ((u32) fp->cmd_vars[0] != 0U) {
+    if (fp->cmd_vars[0] != 0U) {
         fp->mv.mr.SpecialLw.groundVelX =
-            (float) (fp->mv.mr.SpecialLw.groundVelX -
-                     sa_2->speciallw.friction_end);
+            (fp->mv.mr.SpecialLw.groundVelX - sa_2->speciallw.friction_end);
         flt_var += fp->mv.mr.SpecialLw.groundVelX;
         if (flt_var < 0) {
             flt_var = 0;
