+++ src/melee/gm/gmvsmelee.c
@@ -154,8 +154,7 @@ void gm_801A57A8(GameScene* scene_data, VsModeData* vs_data, u8 id)
     if (sss_data->start_game != 0) {
         *vs_data = sss_data->data;
         lbAudioAx_80026F2C(0x18);
-        lbAudioAx_8002702C(8,
-                           lbAudioAx_80026EBC((u16) vs_data->data.rules.xE));
+        lbAudioAx_8002702C(8, lbAudioAx_80026EBC(vs_data->data.rules.xE));
         lbAudioAx_80027168();
     } else {
         gm_SetPendingSceneIndex(id);
